Fix test for custom storage handler
Fixes access issue and results

Testing: ran test
