+++ presto-product-tests/src/test/java/com/facebook/presto/tests/hive/JsonTableDefinition.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.tests.hive;
+
+import com.teradata.tempto.fulfillment.table.TableDefinitionsRepository;
+import com.teradata.tempto.fulfillment.table.hive.HiveDataSource;
+import com.teradata.tempto.fulfillment.table.hive.HiveTableDefinition;
+
+import static com.teradata.tempto.fulfillment.table.hive.InlineDataSource.createResourceDataSource;
+
+public final class JsonTableDefinition
+{
+    private JsonTableDefinition()
+    {
+    }
+
+    @TableDefinitionsRepository.RepositoryTableDefinition
+    public static final HiveTableDefinition SIMPLE_JSON_TABLE = simple_json_table_definition();
+
+    private static HiveTableDefinition simple_json_table_definition()
+    {
+        String tableName = "simple_json_table";
+        HiveDataSource dataSource = createResourceDataSource(tableName, "" + System.currentTimeMillis(), "com/facebook/presto/tests/hive/data/json_data/simple_json_table.json");
+        return HiveTableDefinition.builder()
+                .setName(tableName)
+                .setCreateTableDDLTemplate("" +
+                        "CREATE TABLE %NAME%(" +
+                        "   c_string           string," +
+                        "   c_boolean          boolean," +
+                        "   c_double           double" +
+                        ") " +
+                        "ROW FORMAT SERDE 'org.openx.data.jsonserde.JsonSerDe'" +
+                        "LOCATION '%LOCATION%'")
+                .setDataSource(dataSource)
+                .build();
+    }
+}
